Purge deleted images throws an Exception #133  - Add ability for a BackgroundTask to ensure it is run in a GORM session  - Add transaction for purge deleted images  - Add requiresSession for ImageBackgroundTask, ImageMetadataUpdateBackgroundTask, ScheduleLicenseReMatchAllBackgroundTask, ScheduleMissingImagesBackgroundTask  - Remove UploadFromStagedFileTask as it's no longer used  - Remove LicenseMatchingBackgroundTask for a single image as it was not used and move ScheduleLicenseReMatchAllBackgroundTask to its own file.
