+++ grails-app/views/image/_coreImageMetadataFragment.gsp
@@ -159,7 +159,7 @@
             <g:if test="${isImage}">
                 <button class="btn btn-default" id="btnViewImage" title="View zoomable image"><span class="glyphicon glyphicon-eye-open"> </span></button>
             </g:if>
-            <a class="btn btn-default" href="${createLink(controller:'image', action:'getOriginalFile', id:imageInstance.id, params:[contentDisposition: 'true'])}" title="Download full image" target="imageWindow"><i class="glyphicon glyphicon-download-alt"></i></a>
+            <a class="btn btn-default" href="${createLink(controller:'image', action:'getOriginalFile', id:imageInstance.id, params:[contentDisposition: 'true'])}" title="${g.message(code:'core.image.metadata.download', default: 'Download full image')}" target="imageWindow"><i class="glyphicon glyphicon-download-alt"></i></a>
             <g:if test="${isAdminView}">
                 <button class="btn btn-default" id="btnRegen" title="Regenerate artifacts"><i class="glyphicon glyphicon-refresh"></i></button>
             </g:if>
@@ -169,7 +169,7 @@
                 </button>
             </g:if>
             <g:if test="${!isAdminView && (userId && userId.toString() == imageInstance.uploader) }">
-                <button class="btn btn-danger" id="btnDeleteImage" title="Delete your image)">
+                <button class="btn btn-danger" id="btnDeleteImage" title="${g.message(code:'core.image.metadata.delete', default: 'Delete your image')}">
                     <i class="glyphicon glyphicon-remove  glyphicon-white"></i>
                 </button>
             </g:if>
